+++ src/index.ts
@@ -1,3 +1,3 @@
-export * from "./id";
+export * from "./element_id";
 export { IdList, KnownIdView } from "./id_list";
 export * from "./saved_id_list";
